requirements.txt
# CrimeMap - Crime Data Transparency Platform
# Python 3.10+

# Backend API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
supabase==2.11.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
geoalchemy2==0.15.2

# Ingestion pipeline
httpx==0.28.1             # async HTTP client for fetching reports
beautifulsoup4==4.12.3    # HTML parsing
pdfplumber==0.11.4        # PDF text extraction
openpyxl==3.1.5           # Excel file parsing

# LLM parsing
anthropic==0.40.0         # Claude API for normalization

# Geocoding
googlemaps==4.10.0        # Google Maps Geocoding API

# Task scheduling (for later, but install now)
celery==5.4.0
redis==5.2.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0           # retry logic for API calls
python-dateutil==2.9.0
